+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 # 0.1.1 (Unreleased)
 
+- Add `Altair` consensus structures and pass spec tests [[GH-33](https://github.com/umbracle/eth2-validator/issues/33)]
 - Use multiple accounts in e2e validators for Teky, Prysm and Lighthouse [[GH-25](https://github.com/umbracle/eth2-validator/issues/25)]
 
 # 0.1.0 (12 May, 2022)